refactor(score/): score 패키지 domain으로 이동
