+++ src/main/java/chess/domain/piece/blank/Blank.java
@@ -1,6 +1,6 @@
 package chess.domain.piece.blank;
 
-import chess.score.Score;
+import chess.domain.score.Score;
 import chess.domain.color.Color;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
