+++ src/main/java/chess/domain/state/ChessState.java
@@ -5,8 +5,8 @@ import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
 import chess.domain.position.Position;
 import chess.domain.position.Positions;
-import chess.score.Score;
-import chess.score.ScoreManager;
+import chess.domain.score.Score;
+import chess.domain.score.ScoreManager;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
