+++ src/test/java/chess/domain/score/ScoreTest.java
@@ -2,7 +2,6 @@ package chess.domain.score;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.score.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
